+++ src/core_modules/capture-core-utils/featuresSupport/support.js
@@ -7,6 +7,7 @@ export const FEATURES = Object.freeze({
     newTransferQueryParam: 'newTransferQueryParam',
     exportablePayload: 'exportablePayload',
     changelogs: 'changelogs',
+    trackerImageEndpoint: 'trackerImageEndpoint',
 });
 
 // The first minor version that supports the feature
@@ -16,6 +17,7 @@ const MINOR_VERSION_SUPPORT = Object.freeze({
     [FEATURES.multiText]: 41,
     [FEATURES.customIcons]: 41,
     [FEATURES.exportablePayload]: 41,
+    [FEATURES.trackerImageEndpoint]: 41,
     [FEATURES.newTransferQueryParam]: 41,
     [FEATURES.changelogs]: 41,
 });