+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.container.js
@@ -3,7 +3,7 @@ import React, { useCallback } from 'react';
 import log from 'loglevel';
 import { errorCreator } from 'capture-core-utils';
 // $FlowFixMe
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import {
     useCommonEnrollmentDomainData,
@@ -12,6 +12,9 @@ import {
     updateEnrollmentDate,
     updateIncidentDate,
     showEnrollmentError,
+    updateEnrollmentAndEvents,
+    commitEnrollmentAndEvents,
+    rollbackEnrollmentAndEvents,
 } from '../../common/EnrollmentOverviewDomain';
 import {
     updateEnrollmentDate as updateTopBarEnrollmentDate,
@@ -42,6 +45,7 @@ import {
 export const EnrollmentPageDefault = () => {
     const history = useHistory();
     const dispatch = useDispatch();
+    const { status: widgetEnrollmentStatus } = useSelector(({ widgetEnrollment }) => widgetEnrollment);
     const { enrollmentId, programId, teiId, orgUnitId } = useLocationQuery();
     const { orgUnit, error } = useCoreOrgUnit(orgUnitId);
     const { onLinkedRecordClick } = useLinkedRecordClick();
@@ -125,6 +129,20 @@ export const EnrollmentPageDefault = () => {
     };
 
     const onEnrollmentError = message => dispatch(showEnrollmentError({ message }));
+    const onUpdateEnrollmentStatus = useCallback(
+        (enrollmentToUpdate: Object) => dispatch(updateEnrollmentAndEvents(enrollmentToUpdate)),
+        [dispatch],
+    );
+    const onUpdateEnrollmentStatusError = useCallback(
+        (message) => {
+            dispatch(rollbackEnrollmentAndEvents());
+            dispatch(showEnrollmentError({ message }));
+        },
+        [dispatch],
+    );
+    const onUpdateEnrollmentStatusSuccess = useCallback(() => {
+        dispatch(commitEnrollmentAndEvents());
+    }, [dispatch]);
 
     if (isLoading) {
         return (
@@ -161,7 +179,11 @@ export const EnrollmentPageDefault = () => {
             onUpdateEnrollmentDate={onUpdateEnrollmentDate}
             onUpdateIncidentDate={onUpdateIncidentDate}
             onEnrollmentError={onEnrollmentError}
+            onUpdateEnrollmentStatus={onUpdateEnrollmentStatus}
+            onUpdateEnrollmentStatusSuccess={onUpdateEnrollmentStatusSuccess}
+            onUpdateEnrollmentStatusError={onUpdateEnrollmentStatusError}
             ruleEffects={ruleEffects}
+            widgetEnrollmentStatus={widgetEnrollmentStatus}
             onAccessLostFromTransfer={onAccessLostFromTransfer}
         />
     );