+++ src/core_modules/capture-ui/index.js
@@ -32,3 +32,4 @@ export { IconButton } from './IconButton';
 export { ChevronIcon } from './Icons';
 export { NonBundledIcon } from './NonBundledIcon';
 export { FlatList } from './FlatList';
+export { PreviewImage } from './PreviewImage/PreviewImage.component';